Scripts/Ulduar: Boss Razorscale - Update Boss Razorscale thanks to PrinceCreed - harpoon event needs work - apply trinity coding standard to patch - change minor typo in Boss Ignis Thanks to Lopin
--HG--
branch : trunk

+++ src/server/scripts/Northrend/Ulduar/ulduar/boss_ignis.cpp
@@ -133,7 +133,7 @@ public:
         std::vector<Creature*> triggers;
 
         bool Shattered;
-        uint32 SlagPotGUID;
+        uint64 SlagPotGUID;
         uint32 EncounterTime;
         uint32 ConstructTimer;
